extract outcome_fmt
