+++ core/common/outcome.hpp
@@ -62,19 +62,6 @@ namespace fc::outcome {
   [[noreturn]] inline void raise(T t) {
     raise(make_error_code(t));
   }
-
-  /**
-   * Returns human readable error code string
-   * @param ec - error code
-   * @return string representation of error
-   */
-  inline std::string errorToPrettyString(const std::error_code &ec) {
-    std::stringstream ss;
-    ss << ec.category().name() << " error " << std::to_string(ec.value())
-       << ": \"" << ec.message() + "\"";
-    return ss.str();
-  }
-
 }  // namespace fc::outcome
 
 namespace fmt {
