+++ core/drand/example/drand_example.cpp
@@ -7,7 +7,7 @@
 
 #include "clock/time.hpp"
 #include "common/logger.hpp"
-#include "common/outcome2.hpp"
+#include "common/outcome_fmt.hpp"
 #include "drand/impl/http.hpp"
 
 int main(int argc, char *argv[]) {
