+++ core/node/sync_job.cpp
@@ -7,7 +7,7 @@
 
 #include "common/error_text.hpp"
 #include "common/logger.hpp"
-#include "common/outcome2.hpp"
+#include "common/outcome_fmt.hpp"
 #include "node/blocksync_common.hpp"
 #include "node/chain_store_impl.hpp"
 #include "node/events.hpp"
